+++ src/main/java/hillel/spring/doctor/dto/ReviewOutputDto.java
@@ -7,7 +7,7 @@ public class ReviewOutputDto {
     private Integer id;
     private Integer doctorRecordId;
 
-    private Byte serviceRating;
+    private Integer serviceRating;
     private Byte equipmentRating;
     private Byte qualificationRating;
     private Byte treatmentRating;
